chore(gitignore): exclude generated data and local env files
- /data/* (index.lsif, structures.json, .npy embeddings, snippets.jsonl)
  are all reproducible via tools/download_and_index.sh; committing them
  would balloon the repo with ~130 MB of derived data.
- .claude/ is per-machine harness/IDE config.
- .env / .env.local / frontend/dist should never be committed.
