+++ .gitignore
@@ -143,3 +143,20 @@ cython_debug/
 /setup.py
 
 .DS_Store
+
+# Generated indexing artifacts (regenerated by tools/download_and_index.sh)
+/data/*
+!/data/.keep
+
+# Local IDE / harness config
+.claude/
+
+# Local env overrides (both root and frontend)
+.env
+.env.local
+/frontend/.env
+/frontend/.env.local
+
+# Vite build output
+/frontend/dist/
+
