--- frontend/src/components/Logo.tsx
@@ -1,106 +1,0 @@
-export interface LogoProps extends React.ComponentPropsWithoutRef<"svg"> {
-  size?: number | string;
-}
-
-export function Logo({ size, ...others }: LogoProps) {
-  return (
-    <svg
-      {...others}
-      height={size}
-      viewBox="0 0 238 80"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_1_202)">
-        <path
-          d="M111.648 60.8924C111.648 63.7916 109.298 66.1422 106.399 66.1422H102.81V62.0436H102.682C102.298 62.5984 101.786 63.1748 101.145 63.7722C100.547 64.3275 99.822 64.8399 98.968 65.3097C98.1569 65.779 97.2388 66.1632 96.2142 66.4625C95.232 66.7611 94.2073 66.9102 93.1401 66.9102C90.8346 66.9102 88.7428 66.526 86.864 65.7575C84.9853 64.9464 83.3632 63.8362 81.9968 62.4273C80.6735 60.9758 79.6488 59.2682 78.9228 57.3045C78.1968 55.3401 77.8341 53.1842 77.8341 50.8362C77.8341 48.6588 78.1543 46.5879 78.7947 44.6236C79.4777 42.6173 80.4173 40.8451 81.6126 39.3081C82.8509 37.7714 84.3449 36.5546 86.0955 35.658C87.8462 34.7187 89.8315 34.2491 92.0514 34.2491C94.0583 34.2491 95.9155 34.5693 97.6231 35.2097C99.3737 35.8074 100.804 36.8534 101.914 38.348H102.042V22.9758C102.042 20.0767 104.392 17.7265 107.291 17.7265H111.648V60.8924ZM102.81 50.5795C102.81 48.2315 102.128 46.3318 100.761 44.8798C99.4378 43.4283 97.559 42.7024 95.1254 42.7024C92.6918 42.7024 90.7921 43.4283 89.4257 44.8798C88.1023 46.3318 87.4404 48.2315 87.4404 50.5795C87.4404 52.9281 88.1023 54.8278 89.4257 56.2793C90.7921 57.7312 92.6918 58.4572 95.1254 58.4572C97.559 58.4572 99.4378 57.7312 100.761 56.2793C102.128 54.8278 102.81 52.9281 102.81 50.5795Z"
-          fill="#DC244C"
-        />
-        <path
-          d="M117.863 40.2667C117.863 37.3678 120.214 35.0176 123.113 35.0176H127.47V40.0126H127.598C128.623 38.0916 129.839 36.6613 131.248 35.722C132.657 34.74 134.429 34.2491 136.564 34.2491C137.119 34.2491 137.673 34.2704 138.229 34.3131C138.784 34.3558 139.296 34.4412 139.766 34.5693V43.343C139.082 43.1294 138.399 42.9798 137.717 42.8942C137.076 42.7661 136.393 42.7021 135.667 42.7021C133.831 42.7021 132.38 42.9583 131.312 43.4706C130.245 43.9829 129.412 44.7089 128.815 45.648C128.259 46.5446 127.897 47.6333 127.726 48.9142C127.555 50.1955 127.47 51.6039 127.47 53.1409V60.8921C127.47 63.7913 125.12 66.142 122.22 66.142H117.863V40.2667Z"
-          fill="#DC244C"
-        />
-        <path
-          d="M163.066 62.2362H162.938C161.87 63.9013 160.44 65.0966 158.647 65.8226C156.897 66.548 155.039 66.9108 153.075 66.9108C151.624 66.9108 150.215 66.6976 148.848 66.2709C147.525 65.8866 146.351 65.2887 145.326 64.4777C144.301 63.6661 143.49 62.663 142.892 61.4672C142.295 60.2719 141.996 58.8845 141.996 57.305C141.996 55.5118 142.316 53.9958 142.956 52.758C143.64 51.5197 144.536 50.495 145.646 49.684C146.799 48.8724 148.101 48.2535 149.553 47.8268C151.005 47.3569 152.499 47.0152 154.036 46.8016C155.616 46.5884 157.174 46.4604 158.711 46.4173C160.291 46.3748 161.742 46.3538 163.066 46.3538C163.066 44.6457 162.447 43.3008 161.208 42.3186C160.013 41.294 158.583 40.7816 156.918 40.7816C155.338 40.7816 153.887 41.1234 152.563 41.8063C151.282 42.4472 150.129 43.3438 149.104 44.4961L143.981 39.2446C145.774 37.5798 147.867 36.3416 150.257 35.5304C152.648 34.6765 155.124 34.2496 157.686 34.2496C160.504 34.2496 162.809 34.6125 164.603 35.3383C166.439 36.0214 167.89 37.0461 168.957 38.4121C170.068 39.7785 170.836 41.4651 171.263 43.4714C171.69 45.4357 171.903 47.7197 171.903 50.3239V60.8934C171.903 63.7921 169.553 66.1428 166.654 66.1428H163.066V62.2362ZM160.696 52.5013C159.971 52.5013 159.052 52.5444 157.942 52.6294C156.875 52.6724 155.829 52.843 154.804 53.1417C153.823 53.4409 152.968 53.8892 152.243 54.4866C151.56 55.0845 151.218 55.9171 151.218 56.9848C151.218 58.1375 151.709 58.9911 152.691 59.5459C153.673 60.1013 154.698 60.3785 155.765 60.3785C156.704 60.3785 157.601 60.2504 158.455 59.9942C159.351 59.7381 160.141 59.3753 160.824 58.9055C161.508 58.4362 162.041 57.8383 162.425 57.1123C162.852 56.3869 163.066 55.5328 163.066 54.5512V52.5013H160.696Z"
-          fill="#DC244C"
-        />
-        <path
-          d="M178.572 40.2667C178.572 37.3678 180.922 35.0176 183.821 35.0176H187.794V39.2441H187.922C188.221 38.6467 188.648 38.0489 189.203 37.4512C189.758 36.8534 190.42 36.3198 191.188 35.8501C191.957 35.3805 192.832 34.9962 193.814 34.6974C194.796 34.3985 195.863 34.2491 197.016 34.2491C199.45 34.2491 201.414 34.6333 202.908 35.4018C204.402 36.1277 205.555 37.1523 206.366 38.4756C207.22 39.7995 207.796 41.3575 208.096 43.1507C208.394 44.9438 208.544 46.8866 208.544 48.979V60.8924C208.544 63.7916 206.194 66.1423 203.294 66.1423H198.938V50.8997C198.938 50.0031 198.894 49.0856 198.809 48.1459C198.766 47.1643 198.574 46.2677 198.233 45.4567C197.934 44.6451 197.443 43.9832 196.76 43.4709C196.12 42.9585 195.18 42.7024 193.942 42.7024C192.704 42.7024 191.701 42.9375 190.932 43.4068C190.164 43.8341 189.566 44.4315 189.139 45.2C188.755 45.926 188.499 46.7585 188.371 47.6976C188.243 48.6373 188.178 49.6189 188.178 50.6436V60.8924C188.178 63.7916 185.828 66.1423 182.929 66.1423H178.572V40.2667Z"
-          fill="#DC244C"
-        />
-        <path
-          d="M237.186 37.453C237.186 40.352 234.836 42.7021 231.937 42.7021H228.733V53.0774C228.733 53.931 228.776 54.721 228.861 55.4465C228.946 56.1299 229.139 56.7273 229.437 57.2396C229.736 57.752 230.185 58.1577 230.782 58.4569C231.423 58.7131 232.255 58.8407 233.28 58.8407C233.792 58.8407 234.454 58.7982 235.265 58.7126C236.119 58.5845 236.76 58.3289 237.186 57.9446V62.5241C237.186 64.5793 235.899 66.4911 233.856 66.7178C232.703 66.8459 231.572 66.91 230.462 66.91C228.84 66.91 227.345 66.7394 225.979 66.3976C224.613 66.0564 223.417 65.5226 222.393 64.7971C221.368 64.0286 220.557 63.0465 219.959 61.8507C219.404 60.6554 219.126 59.2039 219.126 57.4963V42.7021H212.978V40.2664C212.978 37.3675 215.329 35.0173 218.228 35.0173H219.126V31.0446C219.126 28.1455 221.477 25.7953 224.376 25.7953H228.733V35.0173H237.186V37.453Z"
-          fill="#DC244C"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M54.4829 73.538L52.8877 29.5569L49.9987 17.9617L69.2821 20.0031V73.2052L57.5026 80.0036L54.4829 73.538Z"
-          fill="#24386C"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M69.2808 20L57.5013 26.8031L33.1926 21.4709L4.73962 33.0556L-0.000854492 20L17.3167 9.99998L34.6396 -1.52588e-05L51.9577 9.99998L69.2808 20Z"
-          fill="#7589BE"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M0 20L11.7795 26.8031L18.6079 47.1107L41.6635 65.558L34.6409 80L17.3181 69.9995L0 59.9995V20Z"
-          fill="#B2BFE8"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M42.4503 54.8871L34.6409 66.4V80.001L45.7222 73.6073L51.4319 65.0814"
-          fill="#24386C"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M34.6457 52.8031L23.5596 33.6089L25.9474 27.2472L35.0251 22.846L45.7222 33.6092L34.6457 52.8031Z"
-          fill="#7589BE"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M23.5596 33.6073L34.6409 40.001V52.7989L24.3921 53.2399L18.1926 45.3239L23.5596 33.6073Z"
-          fill="#B2BFE8"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M34.6409 40L45.7222 33.6068L53.264 46.1643L44.1375 53.705L34.6409 52.7983V40Z"
-          fill="#24386C"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M45.7223 73.6063L57.5018 80.0005V26.8037L46.0687 20.2052L34.6409 13.6068L23.2078 20.2052L11.78 26.8037V53.2021L23.2078 59.8005L34.6409 66.3995L45.7223 60V73.6063ZM45.7223 46.3989L34.6409 52.7979L23.5596 46.3989V33.6063L34.6409 27.2073L45.7223 33.6063V46.3989Z"
-          fill="#DC244C"
-        />
-        <path
-          d="M34.6457 66.4042V52.8011L23.6221 46.4567V60.0373L34.6457 66.4042Z"
-          fill="url(#paint0_linear_1_202)"
-        />
-      </g>
-      <defs>
-        <linearGradient
-          id="paint0_linear_1_202"
-          x1="32.6131"
-          y1="55.4016"
-          x2="21.6284"
-          y2="55.4016"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#FF3364" />
-          <stop offset="1" stopColor="#C91540" stopOpacity="0" />
-        </linearGradient>
-        <clipPath id="clip0_1_202">
-          <rect width="237.186" height="80" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
